Refactor export logic to a unified utility (Closes #22)
